Integrated jwt-cpp library into codebase
jwt-cpp is a third-party C++ library used to process JSON Web Tokens.
We will use this library to decode and validate id tokens sent with
authorized user requests.

Integration is done via a git submodule within the codebase.
A forked version of library repo is mirrored on GCP to support this.

Change-Id: Idaedb2742abb397fc660850140d42a20b5d5432a